move methods from user_repository.py to mongodb_orm.py
